refactor(pages/home): make single-use components
