+++ __tests__/pages/index.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen } from "@testing-library/react";
+import { render } from "@testing-library/react";
 import Index from "../../pages/index";
 
 describe("<Index />", () => {
