+++ components/pages/home/AlternatingFeatures.tsx
