+++ components/pages/home/Hero.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen } from "@testing-library/react";
+import { render } from "@testing-library/react";
 import { Hero } from "./Hero";
 
 describe("<Hero />", () => {
