+++ components/pages/home/LogoCloud.test.tsx
