+++ components/pages/home/LogoCloud.tsx
